+++ src/main/java/com/refinedmods/refinedstorage/mekanism/ChemicalRenderer.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.mekanism.chemical;
+package com.refinedmods.refinedstorage.mekanism;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.BufferBuilder;
@@ -47,25 +47,32 @@ public final class ChemicalRenderer {
                                final int packedRgb,
                                final TextureAtlasSprite sprite) {
         RenderSystem.setShaderTexture(0, sprite.atlasLocation());
+
         final int r = packedRgb >> 16 & 255;
         final int g = packedRgb >> 8 & 255;
         final int b = packedRgb & 255;
+
         final int slotXEnd = x + 16;
         final int slotYEnd = y + 16;
+
         RenderSystem.setShader(GameRenderer::getPositionTexColorShader);
         final Tesselator tesselator = Tesselator.getInstance();
         final BufferBuilder bufferBuilder = tesselator.begin(
             VertexFormat.Mode.QUADS,
             DefaultVertexFormat.POSITION_TEX_COLOR
         );
-        bufferBuilder.addVertex(poseStack.last().pose(), (float) x, (float) slotYEnd, 0.0F)
-            .setUv(sprite.getU0(), sprite.getV1()).setColor(r, g, b, 255);
-        bufferBuilder.addVertex(poseStack.last().pose(), (float) slotXEnd, (float) slotYEnd, 0.0F)
-            .setUv(sprite.getU1(), sprite.getV1()).setColor(r, g, b, 255);
-        bufferBuilder.addVertex(poseStack.last().pose(), (float) slotXEnd, (float) y, 0.0F)
-            .setUv(sprite.getU1(), sprite.getV0()).setColor(r, g, b, 255);
-        bufferBuilder.addVertex(poseStack.last().pose(), (float) x, (float) y, 0.0F)
-            .setUv(sprite.getU0(), sprite.getV0()).setColor(r, g, b, 255);
+        bufferBuilder.addVertex(poseStack.last().pose(), x, slotYEnd, 0)
+            .setUv(sprite.getU0(), sprite.getV1())
+            .setColor(r, g, b, 255);
+        bufferBuilder.addVertex(poseStack.last().pose(), slotXEnd, slotYEnd, 0)
+            .setUv(sprite.getU1(), sprite.getV1())
+            .setColor(r, g, b, 255);
+        bufferBuilder.addVertex(poseStack.last().pose(), slotXEnd, y, 0)
+            .setUv(sprite.getU1(), sprite.getV0())
+            .setColor(r, g, b, 255);
+        bufferBuilder.addVertex(poseStack.last().pose(), x, y, 0)
+            .setUv(sprite.getU0(), sprite.getV0())
+            .setColor(r, g, b, 255);
         BufferUploader.drawWithShader(bufferBuilder.buildOrThrow());
     }