+++ src/main/java/com/refinedmods/refinedstorage/mekanism/ClientModInitializer.java
@@ -1,10 +1,8 @@
 package com.refinedmods.refinedstorage.mekanism;
 
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
-import com.refinedmods.refinedstorage.mekanism.chemical.ChemicalGridInsertionHint;
-import com.refinedmods.refinedstorage.mekanism.chemical.ChemicalResource;
-import com.refinedmods.refinedstorage.mekanism.chemical.ChemicalResourceRendering;
-import com.refinedmods.refinedstorage.mekanism.chemical.ChemicalStorageVariant;
+import com.refinedmods.refinedstorage.mekanism.grid.ChemicalGridInsertionHint;
+import com.refinedmods.refinedstorage.mekanism.storage.ChemicalStorageVariant;
 
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -26,7 +24,7 @@ public final class ClientModInitializer {
         final ResourceLocation diskModel = createMekanismIntegrationIdentifier("block/disk/chemical_disk");
         for (final ChemicalStorageVariant variant : ChemicalStorageVariant.values()) {
             RefinedStorageApi.INSTANCE.getStorageContainerItemHelper().registerDiskModel(
-                Items.INSTANCE.getChemicalStorageDisk(variant),
+                Items.getChemicalStorageDisk(variant),
                 diskModel
             );
         }