+++ src/main/java/com/refinedmods/refinedstorage/mekanism/Items.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.mekanism;
 
-import com.refinedmods.refinedstorage.mekanism.chemical.ChemicalStorageVariant;
+import com.refinedmods.refinedstorage.mekanism.storage.ChemicalStorageVariant;
 
 import java.util.EnumMap;
 import java.util.Map;
@@ -9,29 +9,27 @@ import java.util.function.Supplier;
 import net.minecraft.world.item.Item;
 
 public final class Items {
-    public static final Items INSTANCE = new Items();
-
-    private final Map<ChemicalStorageVariant, Supplier<Item>> chemicalStorageParts =
+    private static final Map<ChemicalStorageVariant, Supplier<Item>> CHEMICAL_STORAGE_PARTS =
         new EnumMap<>(ChemicalStorageVariant.class);
-    private final Map<ChemicalStorageVariant, Supplier<Item>> chemicalStorageDisks =
+    private static final Map<ChemicalStorageVariant, Supplier<Item>> CHEMICAL_STORAGE_DISKS =
         new EnumMap<>(ChemicalStorageVariant.class);
 
     private Items() {
     }
 
-    public Item getChemicalStoragePart(final ChemicalStorageVariant variant) {
-        return chemicalStorageParts.get(variant).get();
+    public static Item getChemicalStoragePart(final ChemicalStorageVariant variant) {
+        return CHEMICAL_STORAGE_PARTS.get(variant).get();
     }
 
-    public void setChemicalStoragePart(final ChemicalStorageVariant variant, final Supplier<Item> supplier) {
-        chemicalStorageParts.put(variant, supplier);
+    public static void setChemicalStoragePart(final ChemicalStorageVariant variant, final Supplier<Item> supplier) {
+        CHEMICAL_STORAGE_PARTS.put(variant, supplier);
     }
 
-    public Item getChemicalStorageDisk(final ChemicalStorageVariant variant) {
-        return chemicalStorageDisks.get(variant).get();
+    public static Item getChemicalStorageDisk(final ChemicalStorageVariant variant) {
+        return CHEMICAL_STORAGE_DISKS.get(variant).get();
     }
 
-    public void setChemicalStorageDisk(final ChemicalStorageVariant variant, final Supplier<Item> supplier) {
-        chemicalStorageDisks.put(variant, supplier);
+    public static void setChemicalStorageDisk(final ChemicalStorageVariant variant, final Supplier<Item> supplier) {
+        CHEMICAL_STORAGE_DISKS.put(variant, supplier);
     }
 }