+++ src/main/java/com/refinedmods/refinedstorage/mekanism/grid/ChemicalGridInsertionHint.java
@@ -1,9 +1,11 @@
-package com.refinedmods.refinedstorage.mekanism.chemical;
+package com.refinedmods.refinedstorage.mekanism.grid;
 
 import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
 import com.refinedmods.refinedstorage.common.api.grid.GridInsertionHint;
 import com.refinedmods.refinedstorage.common.support.tooltip.MouseClientTooltipComponent;
+import com.refinedmods.refinedstorage.mekanism.ChemicalResource;
+import com.refinedmods.refinedstorage.mekanism.ChemicalUtil;
 
 import java.util.Optional;
 
@@ -12,12 +14,12 @@ import mekanism.api.chemical.ChemicalStack;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.world.item.ItemStack;
 
-import static com.refinedmods.refinedstorage.mekanism.chemical.ChemicalResource.ofChemicalStack;
+import static com.refinedmods.refinedstorage.mekanism.ChemicalResource.ofChemicalStack;
 
 public class ChemicalGridInsertionHint implements GridInsertionHint {
     @Override
     public Optional<ClientTooltipComponent> getHint(final ItemStack stack) {
-        return Optional.ofNullable(stack.getCapability(ChemicalResourceType.ITEM_CAPABILITY))
+        return Optional.ofNullable(stack.getCapability(ChemicalUtil.ITEM_CAPABILITY))
             .map(handler -> handler.extractChemical(Long.MAX_VALUE, Action.SIMULATE))
             .filter(result -> !result.isEmpty())
             .map(this::createComponent);